store_service/store_service/requirements.txt
fastapi[all]==0.92.0
uvicorn==0.20.0
gunicorn==20.1.0
prisma==0.8.1
pydantic==1.10.5
pydantic[email]
starlette==0.25.0
python-dotenv==1.0.0
python-jose
passlib
bcrypt

pytest==7.2.2
pytest-asyncio
faker==17.6.0
